+++ tests/AgentAcademy.Server.Tests/PullRequestSyncServiceTests.cs
@@ -145,6 +145,8 @@ public class PullRequestSyncServiceIntegrationTests : IAsyncDisposable
         sc.AddScoped<RoomSnapshotBuilder>();
         sc.AddSingleton<ILogger<WorkspaceRoomService>>(NullLogger<WorkspaceRoomService>.Instance);
         sc.AddScoped<WorkspaceRoomService>();
+
+        sc.AddScoped<IWorkspaceRoomService>(sp => sp.GetRequiredService<WorkspaceRoomService>());
         sc.AddSingleton<ILogger<RoomLifecycleService>>(NullLogger<RoomLifecycleService>.Instance);
         sc.AddScoped<RoomLifecycleService>();
         sc.AddScoped<IRoomLifecycleService>(sp => sp.GetRequiredService<RoomLifecycleService>());